Support for restricting commutativity (#33)
This is one attempt to add support for restricting which types of children are able to commute inside a commutative parent (#5).
For each commutative parent, one can define any number of "children groups", which are lists of children types allowed to commute together.
If no children groups are provided, the current behaviour of letting all children commute remains.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/33
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>