perf: don't clone in `line_merge_and_structured_resolution` (#91)
This is a bigger refactoring as well. I saw the `TODO` and spent a lot of time trying to refactor `cascading_merge` so as to enumerate all its possible outcomes (since roughly speaking, we do know which merges it will return depending on `line_based_merge.has_additional_conflicts`, `full_merge`, and some other booleans). But after that proved to be too convoluted to be practical, I decided to modify `select_best_merge` instead.

I think I did end up succeeding, but I still kept the old function around, since it is used in another context that doesn't care about the line-based merge.

Co-authored-by: Ughur Alakbarov <ugur.alekperov@gmail.com>
Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/91
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>