feat(mergiraf-solve): `--keep-output` option to control the creation of `.orig` files (#165)
Fixes #159

The name is taken from a similar option in `git mergetool`.

The implementation is a little more verbose than I'd like, but that's what the clap devs recommend.[^1]

[^1]: https://github.com/clap-rs/clap/issues/1649#issuecomment-1836240739

It was hard at first to find a way to properly test the thing, but I think I ended up with quite an elegant solution.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/165
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>