fix: Restrict commutativity to avoid issues with comments (#89)
This adds more restrictions on commutative parents in Java, Rust and Go to avoid commutative merging when comments are involved (because comments are not added as types of children that can be restricted).
It will also prevent legitimate merges but I think the bad merges with comments make the whole experience too unreliable.

WIP because I would like to add test cases for the following issues:
* [x] Fixes #26
* [x] Fixes #30
*  ~~[ ] Fixes #63~~
* Presumably also #87, but reproducing it seems complicated

I would also like to do the same sort of work for other languages. It requires going through the grammar to understand what types of children a commutative parent can have and analyze what can go wrong when they swap places.

Co-authored-by: ugur-a <ugur-a@noreply.codeberg.org>
Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/89
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>