Nominate funkeleinhorn to the Developers team (#179)
Hi @funkeleinhorn,

Thanks a lot for your contributions of the TOML and Devicetree support!
In this project we are experimenting with [a governance model](https://codeberg.org/mergiraf/mergiraf/src/branch/main/GOVERNANCE.md), with the aim to proactively grant accesses to contributors and turn this project into a community-owned asset. Whether or not you intend to make other contributions to this project, I'd like to to give you write access by nominating you as a "Developer" in the terms of this model. I totally understand if you'd prefer not to.

I am also happy for feedback on the governance model itself of course :)

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/179
Co-authored-by: wetneb <wetneb@noreply.codeberg.org>
Co-committed-by: wetneb <wetneb@noreply.codeberg.org>