refactor(tree_builder): introduce a type alias to simplify function signatures (#187)
The name is not great... In `extract_conflict_sides`, we have both `successors`, which is a parameter, a multi-map of successors, and `cursor`, which is a PCS node. But in the other methods, the name `cursor` is used pretty consistently to mean exactly what this newly created type alias means. But the type is not used in that many places to matter, so it should be fine.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/187
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>