perf: use rustc-hash in more places (#215)
I made a flamegraph of integration tests, and noticed that a non-trivial amount of time was spent in `Changeset::add`. Digging deeper, it turned out the majority of that had to do with `MultiMap`, because of all the hashing it does. So I took inspiration from #40 and replaced its internal `HashMap`s' hashers with rustc-hash, which indeed reduced the amount of time spent. After that, a couple more bottlenecks became apparent, and I solved them using the same method.

I was testing this with the currently slowest integration test, `java/working/move_and_modify_conflict`, and after all the changes, its runtime went from 0.5s down to 0.4s -- not bad for a simple change like this! The runtime of the whole integration test suite has a greater variance, so it's hard to draw any definitive conclusions from it

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/215
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>