doc: store the list of the supported languages in a separate file and test it (#240)
The motivation for this is two-fold:
- People often forget to update the documentation when adding a new language.
- _I_ forgot to check that the output of `mergiraf languages` was still correct while during #177, which was kindly pointed out at #238. Embarrassing!

It would be nice to do the same for `doc/src/languages.md`, but the list there is intended to be more human-readable, and therefore contains extraneous test that it would be silly to test for. The hope is that if the other file is wrong, then this one is probably wrong as well, so the test above will cover it somewhat.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/240
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>