fix(AttemptsCache::prune): remove the _oldest_ attempts (#298)
When refactoring the method in #287, I didn't notice that the order was
reversed. This PR fixes that and adds a test. The latter required
timeouts, since otherwise the timestamp order was flaky. I initially
went with a small timeout (1 ms), but that turned out to be flaky in CI,
so I increased it to 100 ms

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/298
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>