Revert "trailing_whitespace: call `AstNode::trailing_whitespace` earlier" (#302)
That commit actually introduced a change into method logic. Before, we
first tried to get nodes at the three revisions, and proceeded based on
whether all of those were `Some`. But after the commit, we try to get
the nodes _and_ get their `trailing_whitespace`s, and match on the
success of _that_. I'm not sure if this new behaviour is still okay, but
I'd rather err on the safe side.

This reverts commit 25771fdcf60a6aa30d5801609bb0fc463c51780e, part of
#300

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/302
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>