fix: Spurious indentation changes on untouched nodes (#306)
Fixes #259. Fixes #276.

The issue was indeed caused by the trailing newline included in the doc comments. As a consequence, the whitespace between the doc comment and the function declaration did not contain any newline. When two nodes are separated by only tabs or spaces, Mergiraf considers that the second one starts on the same line as the first one ends, and so that the second node doesn't introduce any indentation shift compared to its parent. This means that all of its contents are then pretty-printed at the indentation level of the parent.

I considered adding some sort of logic to handle this case in the heuristics that add whitespace in the merged text, but given how complex they currently are, I think it wouldn't be a clear improvement. Instead, I am proposing to just strip the last newline of any node, directly when transforming the tree-sitter output into `AstNode`s. It feels simpler to me: in the rest of the algorithm, the characters that belong to a node and those who belong between nodes stay fixed and we don't need to adjust things on the fly.

It has the advantage of letting us remove (in https://codeberg.org/wetneb/mergiraf/commit/12996b797da41ed73290c9b5c1f92167fa1603ba) an on-the-fly trimming logic that I added to handle some cases in Go where newlines are treated as explicit syntax nodes (in import blocs). It turns out that removing that also fixes a test case of commutatively merging XML… so maybe that's not too wrong?

One case where it might be detrimental to strip those newlines might be in multi-line string literals: the inner part of the string literal (without the delimiters) might end with newlines and in those cases they really are an integral part of the node's source. But in that case the newline that we remove will be added again during pretty-printing anyway. And this can also be avoided by turning those newline literals into atomic nodes. And the re-indentation logic for those nodes is already broken, because multi-line string literals probably shouldn't be re-indented at all. I intend to make a follow-up PR to introduce a configuration setting in the `LangProfile` to selectively disable indentation of such nodes.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/306
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>