refactor: return `ParsedMerge` from `line_based_merge` (#323)
The need for rendering and re-parsing the merge has long annoyed me. I'm
pretty sure the reason we were doing it this way is that `ParsedMerge`
wasn't able to outlive `merged_contents` which was defined inside the
function. Solving this did require `String::leak`-ing `merge_contents`,
but what we were doing before was not much better -- in
`cascading_merge`, we'd get `MergeResult` and clone it, which meant we
also cloned the `contents` inside of it.

Now that we got `ParsedMerge` directly, we could avoid the reparse in
two places:
- in the aforementioned `cascading_merge`, we no longer need
`line_based_merge() -> line_based_merge -> (reparse) -> parsed_conflicts
-> resolve_conflict()`
- but notably, in `MergedTree::LineBasedMerge` as well! We can store
`ParsedMerge` there, which means that when we `push_line_based_merge()`
it, the function doesn't need to reparse.

Not everything was quite so easy -- in some methods of `MergedTree`, we
did want to access what was previously `contents`, and one'd think we
would've ended up needing to call `ParsedMerge::render`, but it turns
out that was possible to avoid -- see the newly added
`ParsedMerge::render_conflictless` and its callsites for more
information.

The last 3 commits further make use of the simplification and turn the
`ParsedMerge->MergeResult` transition into a method on the latter. This
is something I wanted to do with `MergedTree->MergeResult` as well, to
show that `MergeResult` is kind of the end point of all the different
merge methods. One could argue that the 3 commits ought to be part of a
separate PR.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/323
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>