perf: make `DfsIterator` exact-sized (#328)
In a couple of places, we collect the result of `AstNode::dfs` into
collections. `Iterator::collect` _may_ use the size hint to pre-allocate
the right amount of storage, so this PR provides that information.

Note that `slice::Iter` created from the cached `dfs` field already was
`ExactSizeIterator`, so no change was necessary there.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/328
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>