perf: avoid calling `Matching::add` in a loop (#329)
It calls `HashMap::insert` internally. One should avoid calling that
method in a loop, and instead pass the iterator into `HashMap::extend`,
so that the map can (hopefully) preallocate enough space.

This PR introduces `Matching::extend` which does exactly that.

The other thing `Matching::add` does is call `Matching::remove`, which I
couldn't find a way to "serialize", so it's still called in the loop.
That can be an area for further improvement.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/329
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>