fix: Don't match nodes of different types (#367)
The pre-matching of nodes based on merged chunks in line-based merges didn't check that the nodes are of the same type.
This made it possible in certain languages to match a list with its first element, for instance (see the test case).
I haven't encountered a merge scenario where this leads to an undesirable output, but discovered this inconsistency while working on something else.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/367
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>