fix: Check that the merged text is syntactically valid and consistent with the merged tree (#368)
Fixes #320.

Instead of just parsing the text obtained in the final rendered merge, this goes a step further and checks that the corresponding trees are "isomorphic" to the merged tree.

The isomorphism check is rather optimistic, as it gives up on sections of the merged tree that are obtained using line-based merges. To be able to also check the validity of those, we'd need to be able to parse snippets of text with a given node type as target, which is an open feature request in tree-sitter: https://github.com/tree-sitter/tree-sitter/issues/711.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/368
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>