tests: Re-run if examples/ directory changed (#381)
This is a workaround for a slightly annoying behaviour that I've been observing for a while: when we add or remove test cases by creating/deleting subdirectories in `examples/`, running `cargo test` again doesn't pick them up automatically.

This is [a known problem of the `rstest` crate](https://github.com/la10736/rstest/issues/256), and the proposed workaround is to use `cargo:rerun-if-changed`.
I don't find it ideal, because it also recompiles the mergiraf binary itself if those files change, which shouldn't be necessary. But it seems to be the best we got so far.
My workaround so far has been to make spurious changes to some rust test file (such as `tests/working.rs`) to trigger a new build. I can also continue to do that…

Intuitively, `Cargo.toml` should let us specify a test resources directory declaratively, without having to resort to a `build.rs`. I have [started a discussion about this](https://internals.rust-lang.org/t/feature-request-specifying-test-resources-in-cargo-toml/22907).

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/381
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>