chore: Remove Ast struct (#395)
So far we have a distinction between `Ast` and `AstNode`, with `Ast` being an entire tree and `AstNode` a node in it. But in practice, the distinction between a tree and its root is pretty vacuous. All the methods of `Ast` just delegate to those on the root, without adding much value. I think the original reason I had this distinction was because the `Ast` was meant to own its source, and then `AstNode`s would borrow from it, but that was before I found out about arenas.

So this PR removes the `Ast` class and replaces all its uses by `AstNode` directly. I have made a slightly convoluted commit log for that, just to make sure only the existing calls to `Ast::root()` are removed, and not the ones to `AstNode::root()` (as those actually do some work). My strategy for that was to make the two methods easier to distinguish by renaming `Ast::root()` to `Ast::redundant_root()` first, then introducing a similar one on `AstNode` (just to make sure replacing all uses of `Ast` by `AstNode` still compiles), and then inlining `AstNode::redundant_root()`.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/395
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>