chore: Refactor `AstNode` creation (#400)
The goal of this PR is to integrate the creation and initialization of the tree-sitter parser inside the constructor of the `AstNode` struct.
This allows for removing a bunch of duplicate initializations of parsers in various areas of the code.
This is in preparation for #5 where we'll need to intertwine the parsing of multiple languages together, making it impossible to use just one tree-sitter parser to parse a given string.

By doing that we are losing the ability to re-use a parser for parsing multiple strings in a row, but I think it should be okay performance-wise.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/400
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>