fix: Preserve initial whitespace before the first child (#408)
Fixes #407. The fix also solves the problem that motivated me to investigate this, namely the missing whitespace in a test case with nested injections.

It turns out our test suite didn't care about leading or trailing whitespace in merge results at all… I have changed that, which required making quite some changes to test cases - mostly overriding the expected output to match the presence or absence of a trailing `\n` character.
It's maybe a bit annoying to be that precise, but I guess we do want to be that precise, don't we?

This also takes the (maybe controversial) approach of ensuring that the root node of each parse spans the whole text to be parsed. The motivation is to be able to encode the leading and trailing whitespace in the root node itself - otherwise to preserve it in the merged output we'd probably need to record it separately. By doing that I also needed to make a special cases for empty files (see comment).
Not particularly proud of this, but I'm tired and out of time for today, so that will be it for now :)

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/408
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>