feat: OCaml support (#426)
This adds effectively two different languages, because there are two different tree-sitter syntaxes for the OCaml implementation and interface files.

Only a very simple case of commutative parents is provided, there might be more to add in the future (though it's unclear because very few things are fully unordered in OCaml).

Co-authored-by: Théophane Hufschmitt <theophane.hufschmitt@routine.co>
Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/426
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: thufschmitt <thufschmitt@noreply.codeberg.org>
Co-committed-by: thufschmitt <thufschmitt@noreply.codeberg.org>