fix: Improve error reporting when working copy is not in a conflict state (#423)
When trying to resolve a conflict with `mergiraf solve` in a working copy which is not in a conflict state, mergiraf reports

```
error while retrieving Base revision for <path>:
git checkout-index: <path> does not exist at stage 1
```

This actually comes directly from `git checkout-index` but is confusing as it's not clear what the issue could be to a user not versed in `checkout-index` speficially (even if proficient with git). The underlying issue is that the index entry is not in a conflict state, and so there is no stage git can recover.

This can be caused by at least two things:

- A new file with a conflict marker is added to the working copy (e.g. maybe work was done in an other working copy then the files were copied over conflict markers and all) (this is the version tested for simplicity)
- The conflict is committed (à la jj I think?) e.g. some merge-adjacent operation was performed by tooling, and in order to report it the tooling just created a commit with the conflict markers (this is the version I hit because of $DAYJOB tooling)

In that case, mergiraf should not even attempt to call `checkout-index`, and should make the error clearer.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/423
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: xmo <xmo@noreply.codeberg.org>
Co-committed-by: xmo <xmo@noreply.codeberg.org>