feat: resolve file revisions from git commits (#434)
If retrieving the revisions from git's index failed, and the conflict markers have full oids, we can use the oids to retrieve the revisions from the repository.

This version has fairly limited applicability as it is quite conservative: it will only trigger on a diff3 conflict marker from which we can extract full oids. It's not entirely clear whether that occurs situations other than `merge-tree` with 3 oids.

We likely do not want to support refnames as those are not stable and the files may have been changed since the original conflict, invalidating the solve, however the constraints could probably be relaxed:

- short oids could be accepted as long as they're resolved uniquely, `-c conflictStyle=diff3 merge-tree oid1 oid2` seems to set a short hash on the base
- if we only have the left and right commits (diff2, there may be diff3 cases where the base is not set?), we could ask git for the `merge-base` and assume that is the base indeed

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/434
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Xavier Morel <xmo@odoo.com>
Co-committed-by: Xavier Morel <xmo@odoo.com>