fix: Check for Jujutsu repo at conflict location (#462)
We warn Jujutsu users when they are using the command `solve`, because
that likely won't work. This is a problem for contributors who use
Jujutsu. For example, tests of the `solve` command might fail, because
Jujutsu is detected.

To prevent this, we check if there is a Jujutsu repository at the
location of the conflicted file that was passed to the `solve` command
(instead of the current working directory). Our tests run in temporary
directories, where no Jujutsu repository will be detected.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/462
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Remo Senekowitsch <remo@buenzli.dev>
Co-committed-by: Remo Senekowitsch <remo@buenzli.dev>