fix(PHP): Don't commutatively merge comments in declaration_lists (#459)
This is calling for a more generic fix, because in theory we should
be restricting all commutative parents to prevent comments from being
commutatively merged. Instead of having to list all types of possible
children explicitly, perhaps we should just forbid the reordering of all
'extra' node types (that tree-sitter will accept to parse anywhere), if
we are able to detect those.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/459
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>