test: reorganize `parsed_merge` tests (#472)
#471 made it clear just how much repetitive code there is in these tests. So I
went on and created a helper to reduce the boilerplate.

Separately, the suite has come to include a large amount of tests for each of
the major methods of `ParsedMerge`: `parse`, `render`, `add_revision_names`,
etc. It was a bit too easy to get lost in the long list of tests, so I put them
into separate modules. Maybe at some point we'll want to split the
`parsed_merge` module itself into multiple files -- moving the relevant tests
would then be as simple as moving the right test module

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/472
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>