refactor(tests): Inline all language-specific parse methods (#530)
I get annoyed of the need to define a new `TestContext::parse_blah` method every time I write a test using a language that's not used in unit tests already. It feels easy enough to provide a filename from which the language can be generated.

So this PR removes them… but it's really a matter of taste, I can also live with adding those methods.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/530
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>