feat: Distinguish between exact and inexact initial matchings (#523)
The tree-matching heuristics we use are often seeded with pre-existing
matchings between the trees to match. This happens in two cases:
1. When we have a line-based merge between the two sources, we create an
   initial matching which pre-matches elements that lie in merged
   regions (as they are identical in all revisions)
2. When we have Base-Left and Base-Right matchings, we compose them to
   obtain an initial Left-Right matching.

So far, those two sorts of initial matchings were handled by the
matching heuristics in the same way. But their nature is quite
different. In the first case, this is a matching of nodes whose sources
are literally the same, so of course they are also isomorphic as
subtrees. On the contrary, in the second case, the initial matching is
obtained by composition of potentially inexact matches. Since the first
pass of the GumTree algorithm is concerned with matching isomorphic
nodes, it is useful for it to know that it can indeed rely on the
knowledge that the initial matches are all exact.

This PR makes it possible to distinguish between the exact and inexact
matches supplied to the matching algorithm. It happens to solve an
existing failing test case (closes #139).

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/523
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>