feat(TypeScript): commutative merging for union and intersection types (#531)
Closes #484.

This is the approach 2. that I mentioned there. I would have preferred
approach 1., but it looks very hard to change the grammar accordingly.
Arguably, not having to change the grammar at all is also a good thing.
The added complexity to the parsing code looks manageable to me.

I don't know if it's feature creep? It doesn't feel that far from
mergiraf's core feature set to me.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/531
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>