chore(changelog): add commit ids to log entries without names (#547)
When we don't squash-merge PRs but rebase them instead, there
is no PR number in the commit line, so it's harder for readers
to inspect the change. It also means git-cliff cannot associate
Codeberg usernames to those commits. This change uses this fact
to take the opportunity to add the commit id, which Codeberg
then turns into a short link when rendering the Markdown.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/547
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>