refactor(merge): `leak` in fewer places (#555)
The first commit is pretty non-controversial. The second does introduce
a possibly weird-looking `Arc<Cow<'static, str>>`, so it could make
sense to run the benchmark suite to see if it's worth it.

It could be possible to avoid requiring `Arc` in a parameter (and to
construct it internally instead), but that unfortunately breaks
`failing.rs`, because it requires passing `contents_*` to two separate
function calls, both times "by value", i.e. as `Cow<'static, str>`. We
could work around this by just cloning in that particular integration
test, but I think it's good in general for
`line_merge_and_structured_resolution` to accept `contents_*`
"by reference", even if that's achieved by an `Arc`

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/555
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>