feat(Git): support reading language from git attributes
When used as a merge driver, `mergiraf` only has the file extension as
guidance for language specification. This may not be suitable in all
cases (e.g., custom languages that are supported-language-adjacent or
`.in` template files that are mostly another language). Allow
specification of the language using the `mergiraf.language` attribute.