governance: Make mathstuf a developer (#566)
Hi @mathstuf,

Thanks a lot for your contributions around C++ extensions, manual language selection and scanning gitattributes :)
In this project we are experimenting with [a governance model](https://codeberg.org/mergiraf/mergiraf/src/branch/main/GOVERNANCE.md), with the aim to proactively grant accesses to contributors and turn this project into a community-owned asset. Whether or not you intend to make other contributions to this project, I'd like to to give you write access by nominating you as a "Developer" in the terms of this model. This will let you review and merge PRs and triage issues. Those accesses will expire if you don't end up using them, so you don't need to worry about them being unnecessary, but I also understand if you prefer not to have them.

I am also happy for feedback on the governance model itself of course :)

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/566
Co-authored-by: wetneb <wetneb@noreply.codeberg.org>
Co-committed-by: wetneb <wetneb@noreply.codeberg.org>