fix(typescript): Bad handling of semicolons in commutative merging of class bodies (#585)
This is a minimized version of a bug reported by Stainless.
The `;` are not recognized as separators in the `class_body`, so the merging algorithm is treating them as independent elements that are being added.
One solution would be to fix the Typescript grammar so that those elements are bundled into the attributes they come at the end of.
One could perhaps come up with other solutions on Mergiraf's side, but it looks like we would need to have to relax the current restriction we have that each separator associated to a commutative group must be equal to the default separator for the commutative parent, up to whitespace.

Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/585
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>