feat: Support for `pyproject.toml` (#582)
This is an attempt at enabling the commutativity of specific fields in a `pyproject.toml` file, following [their specs](https://packaging.python.org/en/latest/specifications/pyproject-toml/).
My impression is that such files are widespread enough to justify having their own language profile, especially because it doesn't cost us much in terms of binary size (as we're reusing a parser that we already include).
Note that I've added the language profile before the TOML one, so that the more specific `pyproject.toml` one is encountered first. Ideally we should not have to worry about this order of definitions and the most specific profile would get selected automatically.

This is prompted by a request from Stainless but I'm doing this pro bono (it doesn't actually solve their motivating conflict which appears in a `tool` section of the `pyproject.toml`).

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/582
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>