chore: Update dependencies (#592)
By running `cargo upgrade -i && cargo update`.

Note the `-i` which I forgot last time, which is necessary for trying
out versions which might be incompatible. This is especially useful for
grammars which tend to be versioned as `0.x.y` and therefore any minor
version increase is treated as a breaking one.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/592
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>