feat(Python): Enable commutation of class attributes, with docstrings (#590)
Closes #589.

I am not very happy with the overall approach, but it feels like it's the best I can do.

Ideally, the bundling of the comments should be done in the grammar, similarly to what we do in Rust, since the attachment is done by Python itself, regardless of the amount of whitespace surrounding it. But I haven't been able to update the grammar accordingly so far.

We could also decide to define the additional comments via a tree-sitter query, but as things stand it would still not remove the need to fork the grammar because the `assignment` nodes are wrapped inside an `expression_statement` node, making it impossible to define the commutative children groups as things stand (except if we also migrate those to support TS queries…)
That being said, there are [good signs that the tree-sitter-python maintainer will accept a move to hide those `expression_statement` nodes](https://github.com/tree-sitter/tree-sitter-python/pull/317) in the upstream grammar. When that gets published, we could then be able to switch back to upstream, assuming we change the `comment_nodes` to use a tree-sitter query instead. But that also has downsides, since that's going to add another query to execute at the beginning of the parsing, complicating the `AstNode` construction a bit further.

In the interest of meeting Stainless' needs, I think it would still be good to ship this (or some variant of it…)

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/590
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>