fix: stack overflow when forcing line-based merges (#623)
Fixes #609.

The reason why this function could recurse infinitely is that it didn't just do a recursion on the `MergedTree`. When encountering an `ExactTree` (which is a leaf, in the sense that it does not contain other `MergedTree`s), it used a convoluted way to recurse into the `ExactTree` by dynamically creating more `ExactTree` instances for the descendants of the AstNode associated with this `ExactTree`.
This is a problem because the class mapping which equates nodes across revisions can equate two versions of a node X, which is a descendent of a node Y in one revision and an ancestor of Y in another. This means that the recursion can jump back to an ancestor of an already-visited node.

Instead, this PR reformulates this recursion at the `AstNode` level (bundled with a revision id, so the function takes a `RevNode`). This guarantees that we recurse of the `AstNode` at a given revision, with the class mapping only used to translate the results of the function to the appropriate `MergedTree` datatype, but not to determine which nodes to recurse on.

In terms of testing, all I have for now is this minimized integration test case, which is not really ideal because it depends on the specifics of how trees are matched. If (when) we change the matching algorithm, it could be that this particular example no longer creates the sort of class mapping that triggers the bug, so the test wouldn't guarantee that we don't reintroduce the bug in the future.

But while debugging this, I have noticed what I think is a bigger problem: the `ClassMapping` seems to violate an important invariant, which is that when mapping an `AstNode` to a `Leader` and then back to an `AstNode` at the same original revision, we should get the same `AstNode`. This seems to fail in certain circumstances, which seems to be a bug in `ClassMapping` itself. I plan to work on a fix for this, and am hopeful that it should be easier to write a narrow unit test for that.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/623
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>