ci: generate release artifacts (#635)
This is another step towards automating the release process: generating all the release artifacts in the CI, via the `cross` tool. It was a bit of a detective work to figure out how to get it to work on Forgejo Actions, so I'm quite happy that this is out of the way ^^
The list of architectures we build artifacts for is stored in `.forgejo/workflows/targets.txt`.
I have tested that they build successfully in #633 (see [this run](https://codeberg.org/mergiraf/mergiraf/actions/runs/1510/jobs/0/attempt/1))

In a follow-up PR I intend to also add a workflow compile and upload the documentation (which should be much easier).
Once that is done, the release process should be as simple as:
- editing Cargo.toml to set a new version
- running `cargo build` for `Cargo.lock` to get updated
- commit
- add a tag of the form "vX.Y.Z"
- push the commit and tag

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/635
Reviewed-by: Ada Alakbarova <ada4a@noreply.codeberg.org>
Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-committed-by: Antonin Delpeuch <antonin@delpeuch.eu>