feat(merge): fallback to `git merge-file` when one of the input files contains conflict markers (#643)
The mechanism is as follows:
1. calculate the conflict regexes when first creating `DisplaySettings`
2. in `main.rs`, use those regexes to look for conflict markers
3. if any are found, fall back to `git merge-file`

The slight problem with that is that since this filtering happens in `main.rs`, the integration tests, most importantly `java/working/leftover_markers`, don't get to benefit from it, and thus fail. One could of course just copy-paste the logic into the `{working,failing}.rs`, but that's less than ideal, as e.g. buggy changes to `main.rs` won't be reflected in the integration tests.

Because of that, I'd like to move that logic into `line_merge_and_structured_resolution`, but that's a bit complicated by the fact that the function returns `MergeResult`, and in particular `contents` -- this means that we'd theoretically need to pipe `git merge-file` to some temporary file, then read the file and write it back to the actual output (be it stdout or another file), which is, again, less than ideal. So maybe the real solution is to refactor `MergeResult` (which is something I've wanted to do for a while) -- I'll see whether I come up with something there. In the meantime, if anyone has better approaches in mind, then do let me know.

Reviewed-on: https://codeberg.org/mergiraf/mergiraf/pulls/643
Reviewed-by: Antonin Delpeuch <wetneb@noreply.codeberg.org>
Co-authored-by: Ada Alakbarova <ada.alakbarova@proton.me>
Co-committed-by: Ada Alakbarova <ada.alakbarova@proton.me>